% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{app1}
\alias{app1}
\title{Data in Appendix 1}
\format{
\subsection{\code{app1}}{

A list with 4 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
app1
}
\description{
Data in Appendix 1 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item McAlpha (used in example 4.5)
\item McBeta (used in example 6.6)
\item McGamma (used in exercise 4.1, example 6.6)
\item McDelta (used in examples 10.4, 10.8, exercise 10.5)
}
}
\keyword{datasets}
