% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom.R
\name{binom}
\alias{binom}
\title{Perform binomial test}
\usage{
binom(
  r,
  n,
  H0 = NULL,
  alternative = c("two.sided", "less", "greater"),
  CI.width = 0.95,
  max.exact.cases = 1e+07,
  do.asymp = FALSE,
  do.exact = TRUE,
  do.CI = TRUE
)
}
\arguments{
\item{r}{Number of "successes"}

\item{n}{Number of trials}

\item{H0}{Null hypothesis value}

\item{alternative}{Type of alternative hypothesis (defaults to \code{two.sided})}

\item{CI.width}{confidence interval width (defaults to \code{0.95})}

\item{max.exact.cases}{Maximum number of cases allowed for exact calculations (defaults to \code{10000000})}

\item{do.asymp}{Boolean for performing/not performing asymptotic calculations (defaults to \code{FALSE})}

\item{do.exact}{Boolean for performing/not performing exact calculations (defaults to \code{TRUE})}

\item{do.CI}{Boolean for creating/not creating confidence interval (defaults to \code{TRUE})}
}
\value{
An ANSMtest object with the results of the binomial test/confidence interval
}
\description{
Perform binomial test
}
