% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch10}
\alias{ch10}
\title{Data used in Chapter 10}
\format{
\subsection{\code{ch10}}{

A list with 26 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch10
}
\description{
Data used in Chapter 10 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item q1 (used in section 10.1.2, examples 10.2, 10.5, 10.9)
\item q2 (used in section 10.1.2, examples 10.2, 10.5, 10.9)
\item death.year (used in examples 10.4, 10.8)
\item diving.rank (used in example 10.10)
\item competitors (used in example 10.10)
\item judges (used in example 10.10)
\item dentistA (used in example 10.11)
\item dentistB (used in example 10.11)
\item questionnaire (used in example 10.12, exercise 10.13)
\item demonstration (used in example 10.12, exercise 10.13)
\item gender (used in exercise 10.13)
\item items (used in example 10.12)
\item ERA (used in exercises 10.1, 10.3, 10.6, 10.7)
\item ESMS (used in exercises 10.1, 10.3, 10.6)
\item SSS (used in exercise 10.7)
\item British (used in example 10.8, exercise 10.10)
\item American (used in example 10.8, exercise 10.10)
\item Canadian (used in example 10.9, exercise 10.10)
\item Australian (used in example 10.9, exercise 10.10)
\item design (used in exercise 10.10)
\item country (used in exercise 10.10)
\item marks (used in exercise 10.11)
\item script (used in exercise 10.11)
\item examiner (used in exercise 10.11)
\item observerA (used in exercise 10.12)
\item observerB (used in exercise 10.12)
}
}
\keyword{datasets}
