% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch11}
\alias{ch11}
\title{Data used in Chapter 11}
\format{
\subsection{\code{ch11}}{

A list with 18 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch11
}
\description{
Data used in Chapter 11 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item parentlimit (used in examples 11.2, 11.3, 11.4, 11.6)
\item reportedtime (used in examples 11.2, 11.3, 11.4, 11.6)
\item age (used in example 11.5)
\item length (used in example 11.5)
\item parentlimit.2 (used in example 11.7)
\item reportedtime.2 (used in example 11.7)
\item days.stored (used in exercise 11.3)
\item rotten (used in exercise 11.3)
\item ERA (used in exercise 11.6)
\item ESMS (used in exercise 11.6)
\item depth (used in exercise 11.8)
\item ammonia (used in exercise 11.8)
\item food.weight.A (used in exercise 11.9)
\item weight.gain.A (used in exercise 11.9)
\item food.weight.B (used in exercise 11.9)
\item weight.gain.B (used in exercise 11.9)
\item SW.England (used in exercise 11.10)
\item N.Scotland (used in exercise 11.10)
}
}
\keyword{datasets}
