% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch12}
\alias{ch12}
\title{Data used in Chapter 12}
\format{
\subsection{\code{ch12}}{

A list with 38 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch12
}
\description{
Data used in Chapter 12 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item feedback.freq (used in example 12.1)
\item PPI.person (used in example 12.1)
\item infection.site (used in examples 12.2, 12.3)
\item district (used in examples 12.2, 12.3)
\item drugYZ (used in example 12.4)
\item side.effect (used in example 12.4)
\item drugAB (used in example 12.5)
\item side.effect.level (used in example 12.5)
\item time.to.failure (used in example 12.6)
\item cause (used in example 12.6)
\item dose (used in examples 12.7, 12.8)
\item dose.side.effect (used in example 12.7, 12.8)
\item platelet.count (used in examples 12.9)
\item spleen.size (used in example 12.9)
\item last.digits (used in example 12.10)
\item accidents (used in example 12.11)
\item accidents.reduced (used in example 12.11)
\item side.effect.new (used in example 12.12)
\item side.effect.old (used in example 12.12)
\item bronchitis (used in exercise 12.1)
\item otitis.media (used in exercise 12.1)
\item welsh.language (used in exercise 12.2)
\item opportunities (used in exercise 12.2)
\item diagnosis (used in exercise 12.3)
\item position.played (used in exercise 12.3)
\item PPI.person.2 (used in exercise 12.4)
\item feedback.satisfaction (used in exercise 12.4)
\item win.opinion (used in exercise 12.5)
\item supporter (used in exercise 12.5)
\item diabetes.status (used in exercise 12.6)
\item ethnic.group (used in exercise 12.6)
\item horse.wins (used in exercise 12.7)
\item F1.wins (used in exercise 12.8)
\item strokes (used in exercise 12.9)
\item recurrent.visits (used in exercise 12.10)
\item holes (used in exercise 12.11)
\item first.response (used in exercise 12.12)
\item second.response (used in exercise 12.12)
}
}
\keyword{datasets}
