% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch13}
\alias{ch13}
\title{Data used in Chapter 13}
\format{
\subsection{\code{ch13}}{

A list with 35 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch13
}
\description{
Data used in Chapter 13 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item physical.activity (used in examples 13.1, 13.2, exercise 13.2)
\item tv.viewing (used in examples 13.1, 13.2, exercise 13.2)
\item gender (used in examples 13.1, 13.2, exercise 13.2)
\item machine (used in example 13.3)
\item output.status (used in example 13.3)
\item material.source (used in example 13.3)
\item drug (used in example 13.4, section 13.2.5)
\item side.effects (used in example 13.4, section 13.2.5)
\item age.group (used in example 13.4, section 13.2.5)
\item dose (used in examples 13.7, 13.8)
\item dose.side.effect (used in examples 13.7, 13.8)
\item alcohol (used in example 13.9)
\item malformation (used in example 13.9)
\item frequency (used in example 13.10)
\item person (used in example 13.10)
\item medicine (used in exercise 13.7, section 13.3.1)
\item response (used in exercise 13.7, section 13.3.1)
\item location (used in exercise 13.7, section 13.3.1)
\item chemo.drug (used in example 13.12)
\item chemo.side.effect (used in example 13.12)
\item group (used in section 13.4)
\item promoted (used in section 13.4)
\item company (used in section 13.4)
\item breakfast.eaten (used in exercise 13.3)
\item VEL (used in exercise 13.3)
\item boys.girls (used in exercise 13.3)
\item cholesterol (used in exercise 13.4)
\item SBP (used in exercise 13.4)
\item schooling (used in exercise 13.5)
\item abortion.attitude (used in exercise 13.5)
\item PPI.ages (used in exercise 13.9)
\item PPI.people (used in exercise 13.9)
\item laid.off (used in exercises 13.10, 13.11)
\item employee.ages (used in exercise 13.10)
\item employee.ages.2 (used in exercise 13.11)
}
}
\keyword{datasets}
