% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch14}
\alias{ch14}
\title{Data used in Chapter 14}
\format{
\subsection{\code{ch14}}{

A list with 3 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch14
}
\description{
Data used in Chapter 14 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item example14.2 (used in examples 14.2, 14.5)
\item X14.4 (used in exercise 14.4)
\item Y14.4 (used in exercise 14.4)
}
}
\keyword{datasets}
