% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch15}
\alias{ch15}
\title{Data used in Chapter 15}
\format{
\subsection{\code{ch15}}{

A list with 11 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch15
}
\description{
Data used in Chapter 15 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item diet (used in section 15.3.5)
\item BMI (used in section 15.3.1)
\item wgt.VLCD (used in section 15.3.2)
\item wgt.norm (used in section 15.3.2)
\item opdiff (used in section 15.3.5)
\item optime.VLCD (used in sections 15.3.3, 15.3.6)
\item optime.norm (used in sections 15.3.3, 15.3.6)
\item los.VLCD (used in section 15.3.6)
\item los.norm (used in section 15.3.6)
\item optime (used in section 15.3.4)
\item los (used in section 15.3.4)
}
}
\keyword{datasets}
