% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch3}
\alias{ch3}
\title{Data used in Chapter 3}
\format{
\subsection{\code{ch3}}{

A list with 25 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch3
}
\description{
Data used in Chapter 3 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item sampleI (used in examples 3.1, 3.2, 3.3, exercise 3.17)
\item sampleII (used in examples 3.1, 3.2, 3.3, exercise 3.17)
\item heartrates1 (used in examples 3.4, 3.11)
\item heartrates2 (used in examples 3.5, 3.6, 3.7)
\item withties (used in example 3.8)
\item tiedifrounded1 (used in example 3.8)
\item tiedifrounded2 (used in example 3.8)
\item ages (used in example 3.8, exercise 3.9)
\item sampleA (used in example 3.12)
\item sampleB (used in examples 3.12, 3.13)
\item sampleA2 (used in example 3.12)
\item sampleA3 (used in example 3.12)
\item heartrates2a (used in example 3.14)
\item heartrates2b (used in example 3.14)
\item sampleIa (used in exercise 3.1)
\item parkingtime (used in exercise 3.3)
\item Svals (used in exercise 3.4)
\item children (used in exercise 3.6)
\item fishlengths (used in exercises 3.7, 3.11)
\item sleeptime (used in exercise 3.10)
\item weightloss (used in exercise 3.12)
\item plants (used in exercise 3.13)
\item birthprops (used in exercise 3.14)
\item assembly (used in exercise 3.15)
\item weightchange (used in exercise 3.16)
}
}
\keyword{datasets}
