% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch3data}
\alias{ch3data}
\title{Data used in Chapter 3}
\format{
\subsection{\code{ch3data}}{

A list with 25 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch3data
}
\description{
Data used in Chapter 3 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item sampleI (Examples 3.1, 3.2, 3.3, Exercise 3.17)
\item sampleII (Examples 3.1, 3.2, 3.3, Exercise 3.17)
\item heartrates1 (Examples 3.4, 3.11)
\item heartrates2 (Examples 3.5, 3.6, 3.7)
\item withties (Example 3.8)
\item tiesifrounded1 (Example 3.8)
\item tiesifrounded2 (Example 3.8)
\item ages (Example 3.9)
\item sampleA (Examples 3.12, 3.13)
\item sampleB (Examples 3.12, 3.13)
\item sampleA2 (amended version of sampleA, used in example 3.12)
\item sampleA3 (amended version of sampleA, used in example 3.12)
\item heartrates2a (amended version of heartrates2, used in example 3.14)
\item heartrates2b (amended version of heartrates2, used in example 3.14)
\item sampleIa (amended version of sampleI, used in exercise 3.1)
\item parkingtime (Exercise 3.3)
\item Svals (Exercise 3.4)
\item children (Exercise 3.6)
\item fishlengths (Exercises 3.7, 3.11)
\item sleeptime (Exercise 3.10)
\item weightloss (Exercise 3.12)
\item plants (Exercise 3.13)
\item birthprops (Exercise 3.14)
\item assembly (Exercise 3.15)
\item weightchange (Exercise 3.16)
}
}
\keyword{datasets}
