% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch4}
\alias{ch4}
\title{Data used in Chapter 4}
\format{
\subsection{\code{ch4}}{

A list with 20 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch4
}
\description{
Data used in Chapter 4 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item breaks (used in example 4.2)
\item ages (used in example 4.4)
\item precipitation (used in example 4.13)
\item tosses1 (used in example 4.14)
\item tosses2 (used in example 4.14)
\item tosses3 (used in example 4.14)
\item births (used in example 4.15)
\item times.as.degrees (used in example 4.16)
\item dates.as.degrees (used in example 4.17)
\item waiting.time (used in exercise 4.2)
\item visiting.supporters (used in exercise 4.3)
\item days.waiting (used in exercise 4.8)
\item rainfall.by.latitude (used in exercise 4.9)
\item points (used in exercise 4.10)
\item rainfall.DRC (used in exercise 4.11)
\item piped.water.DRC (used in exercise 4.12)
\item accident.bearings (used in exercise 4.13)
\item board.angles (used in exercise 4.14)
\item arrow.angles (used in exercise 4.15)
\item football.results (used in exercise 4.17)
}
}
\keyword{datasets}
