% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch5}
\alias{ch5}
\title{Data used in Chapter 5}
\format{
\subsection{\code{ch5}}{

A list with 21 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch5
}
\description{
Data used in Chapter 5 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item LVF (used in example 5.1, exercise 6.2)
\item RVF (used in example 5.1, exercise 6.2)
\item arithmetic (used in example 5.2)
\item bp (used in example 5.3)
\item bp.incorrect (used in example 5.3)
\item yr0910 (used in example 5.10)
\item yr1314 (used in example 5.10)
\item bp.diff (used in exercise 5.1)
\item LabI (used in exercise 5.2)
\item LabII (used in exercise 5.2)
\item parent (used in exercise 5.4)
\item online (used in exercise 5.5)
\item lectures (used in exercise 5.5)
\item additiveA (used in exercise 5.9)
\item additiveB (used in exercise 5.9)
\item round2 (used in exercise 5.10)
\item round3 (used in exercise 5.10)
\item pollA (used in exercise 5.11)
\item pollB (used in exercise 5.11)
\item kHz0.125 (used in exercise 5.12)
\item kHz0.25 (used in exercise 5.12)
}
}
\keyword{datasets}
