% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch6}
\alias{ch6}
\title{Data used in Chapter 6}
\format{
\subsection{\code{ch6}}{

A list with 41 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch6
}
\description{
Data used in Chapter 6 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item groupA (used in examples 6.1, 6.2, 6.3, 6.10, 6.17)
\item groupB (used in examples 6.1, 6.2, 6.3, 6.10, 6.17)
\item groupA.sch2 (used in example 6.4)
\item groupB.sch2 (used in example 6.4)
\item groupA.sch2.grp (used in example 6.5)
\item groupB.sch2.grp (used in example 6.5)
\item McGamma (used in example 6.6)
\item McBeta (used in example 6.6)
\item males (used in examples 6.7, 6.8)
\item females (used in examples 6.7, 6.8)
\item sampleI (used in example 6.9)
\item sampleII (used in example 6.9)
\item typeA (used in examples 6.11, 6.12, 6.13, exercises 6.11, 6.12)
\item typeB (used in examples 6.11, 6.12, 6.13, exercises 6.11, 6.12)
\item groupI (used in example 6.14)
\item groupII (used in example 6.14)
\item groupI.trimmed (used in example 6.14)
\item groupI.amended (used in example 6.14)
\item salivaF (used in examples 6.15, 6.16)
\item salivaM (used in examples 6.15, 6.16)
\item sex (used in example 6.18)
\item temp.H (used in exercise 6.1)
\item temp.L (used in exercise 6.1)
\item DMF.M (used in exercise 6.3)
\item DMF.F (used in exercise 6.3)
\item weight.diabetic (used in exercise 6.4)
\item weight.normal (used in exercise 6.4)
\item cooling.time.standard (used in exercise 6.5)
\item cooling.time.cheap (used in exercise 6.5)
\item wait.1979 (used in exercise 6.6)
\item wait.1983 (used in exercise 6.6)
\item activity.boys (used in exercise 6.7)
\item activity.girls (used in exercise 6.7)
\item time.withoutLD (used in exercises 6.13, 6.14)
\item time.withLD (used in exercises 6.13, 6.14)
\item doseI (used in exercise 6.15)
\item doseII (used in exercise 6.15)
\item doseI.2 (used in exercise 6.15)
\item travel (used in exercise 6.16)
\item politics (used in exercise 6.16)
\item twins (used in exercise 6.17)
}
}
\keyword{datasets}
