% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch7}
\alias{ch7}
\title{Data used in Chapter 7}
\format{
\subsection{\code{ch7}}{

A list with 47 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch7
}
\description{
Data used in Chapter 7 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item affordability (used in example 7.1, exercise 7.16)
\item regions (used in example 7.1, exercise 7.16)
\item age (used in example 7.2)
\item positions (used in example 7.2)
\item dementia.age (used in examples 7.3, 7.9)
\item features (used in examples 7.3, 7.9)
\item time (used in examples 7.4, 7.5)
\item surgeon (used in examples 7.4, 7.5)
\item pulse (used in example 7.6)
\item student (used in example 7.6)
\item time.period (used in example 7.6)
\item nodes (used in example 7.7)
\item treatment (used in example 7.7)
\item block (used in example 7.7)
\item outcome (used in example 7.8)
\item member (used in example 7.8)
\item climb (used in example 7.8)
\item procedure.time (used in example 7.10)
\item team.member (used in example 7.10)
\item sentences (used in exercise 7.2)
\item author (used in exercise 7.2)
\item head.width (used in exercise 7.4)
\item species (used in exercise 7.4)
\item braking.distance (used in exercise 7.5)
\item speed (used in exercise 7.5)
\item platelet.count (used in exercise 7.6)
\item spleen.size (used in exercise 7.6)
\item liver.weight (used in exercise 7.7)
\item dose (used in exercise 7.7)
\item house (used in exercise 7.7)
\item mark (used in exercise 7.8)
\item scheme (used in exercise 7.8)
\item candidate (used in exercise 7.8)
\item prem.contractions (used in exercise 7.9)
\item drug (used in exercise 7.9)
\item patient (used in exercise 7.9)
\item births (used in exercise 7.11)
\item week (used in exercise 7.11)
\item weekday (used in exercise 7.11)
\item names.recalled (used in exercise 7.12)
\item group (used in exercise 7.12)
\item medical.student (used in exercise 7.12)
\item soc.media.use (used in exercise 7.14)
\item participant (used in exercise 7.14)
\item day (used in exercise 7.14)
\item braking.distance.2 (used in exercise 7.15)
\item initial.speed (used in exercise 7.15)
}
}
\keyword{datasets}
