% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch8}
\alias{ch8}
\title{Data used in Chapter 8}
\format{
\subsection{\code{ch8}}{

A list with 29 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch8
}
\description{
Data used in Chapter 8 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item plant.weight (used in example 8.2)
\item growth.hormone (used in examples 8.6, 8.7)
\item undersoil.heating (used in examples 8.6, 8.7)
\item plant.weight.2 (used in example 8.6)
\item plant.weight.3 (used in examples 8.4, 8.5)
\item plant.weight.4 (used in example 8.7)
\item sequence (used in example 8.9)
\item periodI (used in example 8.9)
\item periodII (used in example 8.9)
\item sentences (used in example 8.10)
\item authors (used in example 8.10)
\item prey.preference (used in example 8.11)
\item prey (used in example 8.11)
\item larva (used in example 8.11)
\item game.time (used in exercise 8.3)
\item experience (used in exercise 8.3)
\item game (used in exercise 8.3)
\item periodI.mistakes.AB (used in exercise 8.6)
\item periodII.mistakes.AB (used in exercise 8.6)
\item periodI.mistakes.BA (used in exercise 8.6)
\item periodII.mistakes.BA (used in exercise 8.6)
\item periodI.time.AB (used in exercise 8.7)
\item periodII.time.AB (used in exercise 8.7)
\item periodI.time.BA (used in exercise 8.7)
\item periodII.time.BA (used in exercise 8.7)
\item seizure.score (used in exercises 8.8, 8.9)
\item hospital (used in exercises 8.8, 8.9)
\item silver.content (used in exercise 8.10)
\item dynasty (used in exercise 8.10)
}
}
\keyword{datasets}
