% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ch9}
\alias{ch9}
\title{Data used in Chapter 9}
\format{
\subsection{\code{ch9}}{

A list with 23 data vectors
}
}
\source{
"Applied Nonparametric Statistical Methods" (5th edition)
}
\usage{
ch9
}
\description{
Data used in Chapter 9 of "Applied Nonparametric Statistical Methods" (5th edition)
\itemize{
\item symp.survtime (used in examples 9.1, 9.3)
\item symp.censor (used in examples 9.1, 9.3)
\item asymp.survtime (used in examples 9.1, 9.3)
\item asymp.censor (used in examples 9.1, 9.3)
\item sampleI.survtime (used in following example 9.3, example 9.4)
\item sampleI.censor (used in example 9.4)
\item sampleII.survtime (used in example 9.4)
\item sampleII.survtime.2 (used in following example 9.3)
\item sampleII.censor (used in example 9.4)
\item samplesAB.survtime (used in example 9.6)
\item samplesAB.censor (used in example 9.6)
\item samplesAB (used in example 9.6)
\item samplesXYZ.survtime (used in example 9.7)
\item samplesXYZ.censor (used in example 9.7)
\item samplesXYZ (used in example 9.7)
\item boys.toothtime (used in exercise 9.2)
\item girls.toothtime (used in exercise 9.2)
\item regimeA.survtime (used in exercises 9.5, 9.6)
\item regimeA.censor (used in exercises 9.5, 9.6)
\item regimeB.survtime (used in exercises 9.5, 9.6)
\item regimeB.censor (used in exercises 9.5, 9.6)
\item bulbA (used in exercise 9.8)
\item bulbB (used in exercise 9.8)
}
}
\keyword{datasets}
