% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer.von.mises.R
\name{cramer.von.mises}
\alias{cramer.von.mises}
\title{Perform Cramer-von Mises test}
\usage{
cramer.von.mises(x, y, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}

\item{alternative}{Type of alternative hypothesis (defaults to \code{two.sided})}
}
\value{
An ANSMtest object with the results from applying the function
}
\description{
\code{cramer.von.mises()} performs the Cramer-von Mises test and is used in chapter 6 of "Applied Nonparametric Statistical Methods" (5th edition)
}
\examples{
# Example 6.16 from "Applied Nonparametric Statistical Methods" (5th edition)
cramer.von.mises(ch6$salivaF, ch6$salivaM)
cramer.von.mises(ch6$salivaF, ch6$salivaM, alternative = "greater")

}
