% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mantel.haenszel.R
\name{mantel.haenszel}
\alias{mantel.haenszel}
\title{Perform Mantel-Haenszel test}
\usage{
mantel.haenszel(x, y, z, do.asymp = TRUE)
}
\arguments{
\item{x}{Binary factor of same length as y, z}

\item{y}{Binary factor of same length as x, z}

\item{z}{Factor of same length as x, y}

\item{do.asymp}{Boolean indicating whether or not to perform asymptotic calculations (defaults to \code{TRUE})}
}
\value{
An ANSMtest object with the results from applying the function
}
\description{
\code{mantel.haenszel()} performs the Mantel-Haenszel test and is used in chapter 13 of "Applied Nonparametric Statistical Methods" (5th edition)
}
\examples{
# Example 13.4 from "Applied Nonparametric Statistical Methods" (5th edition)
mantel.haenszel(ch13$drug, ch13$side.effects, ch13$age.group)

#  from "Applied Nonparametric Statistical Methods" (5th edition)

}
