% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ANSMstat.R
\name{print.ANSMstat}
\alias{print.ANSMstat}
\title{Prints an ANSMstat object}
\usage{
\method{print}{ANSMstat}(x, ...)
}
\arguments{
\item{x}{An ANSMstat object}

\item{...}{Further arguments relevant to the default \code{print} function}
}
\value{
No return value, called to display results
}
\description{
\code{print.ANSMstat()} prints the output contained in an ANSMstat object
}
