% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ANSMtest.R
\name{print.ANSMtest}
\alias{print.ANSMtest}
\title{Prints an ANSMtest object}
\usage{
\method{print}{ANSMtest}(x, ...)
}
\arguments{
\item{x}{An ANSMtest object}

\item{...}{Further arguments relevant to the default \code{print} function}
}
\value{
No return value, called to display results
}
\description{
\code{print.ANSMtest()} prints the output contained in an ANSMtest object
}
