% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.test.R
\name{range.test}
\alias{range.test}
\title{Perform Range test}
\usage{
\method{range}{test}(x, alternative = c("two.sided"), minx = 0, maxx = 360)
}
\arguments{
\item{x}{Numeric vector}

\item{alternative}{Type of alternative hypothesis (defaults to \code{c("two.sided")})}

\item{minx}{Minimum value for x (defaults to \code{0})}

\item{maxx}{Maximum value for x (defaults to \code{360})}
}
\value{
An ANSMtest object with the results from applying the function
}
\description{
\code{range.test()} performs the Range test and is used in chapter 4 of \verb{Applied Nonparametric Statistical Methods} (5th edition)
}
\examples{
# Example 4.17 from `Applied Nonparametric Statistical Methods` (5th edition)
#range.test(ch4$dates.as.degrees)

# Exercise 4.13 from `Applied Nonparametric Statistical Methods` (5th edition)
#range.test(ch4$accident.bearings)

}
