% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapiro.test.ANSM.R
\name{shapiro.test.ANSM}
\alias{shapiro.test.ANSM}
\title{Perform Shapiro-Wilk test of Normality}
\usage{
shapiro.test.ANSM(x, alternative = c("two.sided"))
}
\arguments{
\item{x}{Numeric vector}

\item{alternative}{Type of alternative hypothesis (defaults to \code{c("two.sided")})}
}
\value{
An ANSMtest object with the results from applying the function
}
\description{
\code{shapiro.test.ANSM()} is a wrapper for stats::shapiro.test - performs the Shapiro-Wilk test of Normality and is used in chapters 4 and 5 of \verb{Applied Nonparametric Statistical Methods} (5th edition)
}
\examples{
# Example 4.4 from `Applied Nonparametric Statistical Methods` (5th edition)
shapiro.test.ANSM(ch4$ages)

# Example 5.3 from `Applied Nonparametric Statistical Methods` (5th edition)
shapiro.test.ANSM(ch5$bp.incorrect)

}
